Variable Renamed
isOn was changed to cameraActive. The new name provides a clearer description that this boolean variable indicates whether the camera is currently active or not.

videoCapture was renamed to videoStream. The name videoStream better reflects its role, i.e., it's capturing a stream of video frames, not just a single capture.

threadLoop was renamed to runLoop. This new name describes more precisely its function — running a loop that is specifically intended for continuously processing video frames.

postFrame was changed to deliverFrame. This new name makes it clear that the purpose of this function is to deliver (i.e., post) frames to another component or function for further processing.

start and stop were renamed to startCapture and endCapture, respectively. The new names are more descriptive and indicate precisely what these functions do, i.e., start capturing video and end capturing video.

deviceID and apiID were renamed to deviceIndex and apiIndex. The new names are clearer and indicate that these are indices (or identifiers) of the device and API used for video capture.

The SceneCallback interface was renamed to FrameCallback, the nextScene function was renamed to onNewFrame, and the sceneCallback instance variable was renamed to frameCallback. These new names clearly state that this callback interface, its method, and the instance variable are all about processing new frames, not "scenes".